This is an ip core that supports rk4 diff equations.
There are two provided versions of the diff solver -- Floating Point and
Integer.

The floating point uses 32 bit IEEE floating point standard to represent
the input and output. It is easier to interact because the floating
point can represent a large range of values and it is an
industry-standard format. Since it is only 32 bits, it has less
precision than the 64 bit integer system. ( When adding a very small
value with a large value, the small value diminishes because it can not
be represented in the larger values' exponent, such as adding 10 with
.000000001 would result in a 10) It is also more difficult to meet
timing because it instantiates more FP adders and multipliers. (18,000
ns setup time violation)

The integer system uses 64 bit integer with a customized exponent to
represent the input and output. It still cannot meet the 200 MHz clock
requirement. (7,700 ns setup time violation). It has better precision
than the 32 bit floating point system. However, it does not protect the
user from overflow. (If the input is already the maximum value, the
output cannot be greater than that because the input and output shares
the same exponent) It is also more difficult to interact with. Please
see the comment in solver_rk4_fp.sv for more detailed explanations.